# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API for mobile app <br/> <br/> Run `php artisan l5-swagger:generate` to generate updated API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/sensor-type/list": {
            "get": {
                "tags": [
                    "Sensor Type"
                ],
                "summary": "List of sensor types",
                "operationId": "sensorTypeList",
                "responses": {
                    "200": {
                        "description": "List of sensors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SensorType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/sensor-unit/list": {
            "get": {
                "tags": [
                    "Sensor Unit"
                ],
                "summary": "List of sensor units",
                "operationId": "sensorUnitList",
                "responses": {
                    "200": {
                        "description": "List of sensors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SensorUnitResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sensor-unit/{sensorUnit}/sensor-type/list": {
            "get": {
                "tags": [
                    "Sensor Unit"
                ],
                "summary": "List of sensor types for a sensor unit",
                "operationId": "sensorUnitSensorTypeList",
                "parameters": [
                    {
                        "name": "sensorUnit",
                        "in": "path",
                        "description": "Sensor unit ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sensor types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SensorType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SensorType": {
                "title": "Sensor Type",
                "description": "Represents a sensor type.",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the sensor type.",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the sensor type.",
                        "type": "string"
                    },
                    "slug": {
                        "description": "The slug of the sensor type.",
                        "type": "string"
                    },
                    "description": {
                        "description": "A description of the sensor type.",
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "description": "The status of the sensor type.",
                        "type": "string",
                        "default": "active",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "avg": {
                        "description": "The average value of last 24hrs.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SensorUnitResource": {
                "title": "Sensor Unit Resource",
                "required": [
                    "id",
                    "name",
                    "serial_number",
                    "slug",
                    "description",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the resource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "The name of the resource",
                        "type": "string"
                    },
                    "serial_number": {
                        "description": "The serial number of the resource",
                        "type": "string"
                    },
                    "slug": {
                        "description": "The slug of the resource",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the resource",
                        "type": "string"
                    },
                    "status": {
                        "description": "The status of the resource",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}